telegram_tiktok/requirements.txt
aiogram>=3.4.0
aiohttp>=3.9.3
aiohttp-socks>=0.8.4
python-dotenv>=1.0.1
